// types/enhanced-chat.ts
export interface MessageReaction {
  id: string
  message_id: string
  user_id: string
  emoji: string
  created_at: string
  user?: {
    name: string
    profilepicture?: string
  }
}